# optimize: select virtual services