--- a/pilot/pkg/model/virtualservice.go
+++ b/pilot/pkg/model/virtualservice.go
@@ -37,7 +37,7 @@ func SelectVirtualServices(vsidx virtualServiceIndex, configNamespace string, ho
 	vsset := sets.New[types.NamespacedName]()
 
 	addVirtualService := func(vs config.Config, hosts hostClassification) {
-		key := types.NamespacedName{Namespace: vs.Namespace, Name: vs.Name}
+		key := vs.NamespacedName()
 
 		if vsset.Contains(key) {
 			return
@@ -191,7 +191,7 @@ func mergeVirtualServicesIfNeeded(
 		rule := vs.Spec.(*networking.VirtualService)
 		// it is delegate, add it to the indexer cache along with the exportTo for the delegate
 		if len(rule.Hosts) == 0 {
-			delegatesMap[config.NamespacedName(vs)] = vs
+			delegatesMap[vs.NamespacedName()] = vs
 			var exportToSet sets.Set[visibility.Instance]
 			if len(rule.ExportTo) == 0 {
 				// No exportTo in virtualService. Use the global default
@@ -213,7 +213,7 @@ func mergeVirtualServicesIfNeeded(
 					}
 				}
 			}
-			delegatesExportToMap[config.NamespacedName(vs)] = exportToSet
+			delegatesExportToMap[vs.NamespacedName()] = exportToSet
 
 			continue
 		}