--- a/pilot/pkg/serviceregistry/serviceentry/util.go
+++ b/pilot/pkg/serviceregistry/serviceentry/util.go
@@ -27,7 +27,7 @@ func getWorkloadServiceEntries(ses []config.Config, wle *networking.WorkloadEntr
 	for i, cfg := range ses {
 		se := cfg.Spec.(*networking.ServiceEntry)
 		if se.WorkloadSelector != nil && labels.Instance(se.WorkloadSelector.Labels).Match(wle.Labels) {
-			out[config.NamespacedName(cfg)] = &ses[i]
+			out[cfg.NamespacedName()] = &ses[i]
 		}
 	}