**Please provide a description of this PR:**
Optimizations to SelectVirtualServices which is a very hot method:
- use `types.NamespacedName` as `set` type to avoid string concatenation and allocations
- move `vsset.Contains` to the top, to prevent multiple checks for each non-matched rule
- use `return` instead of `break` to break both loops instead of only one