@sschepens 2023-10-05
Benchmark resulsts:
```
name                                              old time/op    new time/op    delta
ConvertIstioListenerToWrapper/small-exact-8         2.43µs ± 3%    2.18µs ± 5%  -10.46%  (p=0.000 n=9+9)
ConvertIstioListenerToWrapper/small-wildcard-8      1.42µs ± 6%    1.21µs ± 5%  -14.60%  (p=0.000 n=9+9)
ConvertIstioListenerToWrapper/small-match-all-8     4.21µs ± 5%    4.17µs ±10%     ~     (p=0.549 n=9+10)
ConvertIstioListenerToWrapper/middle-exact-8        26.6µs ± 3%    25.7µs ± 7%   -3.44%  (p=0.006 n=8+9)
ConvertIstioListenerToWrapper/middle-wildcard-8     25.0µs ± 5%    22.3µs ± 2%  -10.48%  (p=0.000 n=9+9)
ConvertIstioListenerToWrapper/middle-match-all-8    48.3µs ± 7%    50.7µs ±11%     ~     (p=0.105 n=10+10)
ConvertIstioListenerToWrapper/big-exact-8            103µs ± 6%     106µs ± 6%   +3.40%  (p=0.035 n=10+10)
ConvertIstioListenerToWrapper/big-wildcard-8         101µs ± 9%      91µs ± 2%   -9.91%  (p=0.000 n=9+10)
ConvertIstioListenerToWrapper/big-match-all-8        135µs ± 3%     144µs ±10%   +6.06%  (p=0.005 n=10+10)

name                                              old alloc/op   new alloc/op   delta
ConvertIstioListenerToWrapper/small-exact-8         2.64kB ± 0%    2.40kB ± 0%   -9.09%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/small-wildcard-8        640B ± 0%      400B ± 0%  -37.50%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/small-match-all-8     9.74kB ± 0%    9.79kB ± 0%   +0.52%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/middle-exact-8        13.9kB ± 0%    11.8kB ± 0%  -15.16%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/middle-wildcard-8     4.22kB ± 0%    1.82kB ± 0%  -56.82%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/middle-match-all-8     102kB ± 0%     109kB ± 0%   +6.29%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/big-exact-8           23.1kB ± 0%    16.7kB ± 0%  -27.38%  (p=0.000 n=9+10)
ConvertIstioListenerToWrapper/big-wildcard-8        11.0kB ± 0%     3.8kB ± 0%  -65.60%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/big-match-all-8        348kB ± 0%     359kB ± 0%   +3.22%  (p=0.000 n=10+8)

name                                              old allocs/op  new allocs/op  delta
ConvertIstioListenerToWrapper/small-exact-8           29.0 ± 0%      19.0 ± 0%  -34.48%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/small-wildcard-8        19.0 ± 0%       9.0 ± 0%  -52.63%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/small-match-all-8       32.0 ± 0%      22.0 ± 0%  -31.25%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/middle-exact-8           137 ± 0%        37 ± 0%  -72.99%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/middle-wildcard-8        118 ± 0%        18 ± 0%  -84.75%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/middle-match-all-8       146 ± 0%        47 ± 0%  -67.81%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/big-exact-8              345 ± 0%        45 ± 0%  -86.96%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/big-wildcard-8           323 ± 0%        23 ± 0%  -92.88%  (p=0.000 n=10+10)
ConvertIstioListenerToWrapper/big-match-all-8          357 ± 0%        60 ± 0%  -83.19%  (p=0.000 n=10+10)
```