@howardjohn 2023-10-05
This version is fine, lgtm. thanks!

On Thu, Oct 5, 2023 at 11:24 AM sschepens ***@***.***> wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In pilot/pkg/model/virtualservice.go
> <https://github.com/istio/istio/pull/47197#discussion_r1347818590>:
>
> >  	addVirtualService := func(vs config.Config, hosts hostClassification) {
> -		vsname := vs.Name + "/" + vs.Namespace
> +		key := types.NamespacedName{Namespace: vs.Namespace, Name: vs.Name}
>
> so, do you prefer making it generic? generics have a slight performance
> impact, that's probably what you're seeing.
>
> i just pushed a version where I created a NamespacedName() method for
> Config struct to use that instead of the generic method, i can revert
> that if u want.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/istio/istio/pull/47197#discussion_r1347818590>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAEYGXKQL7HYSCH7W6ZOCRTX533OVAVCNFSM6AAAAAA5UKF5C2VHI2DSMVQWIX3LMV43YUDVNRWFEZLROVSXG5CSMV3GSZLXHMYTMNRQGQ3DOMRZGY>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>